\name{alpha}
\alias{alpha}
\title{Calculate alpha function}
\usage{
  alpha(n, k, m)
}
\arguments{
  \item{n}{the total number of individuals in the sample}

  \item{m}{the number of individuals to interpolate S for.}

  \item{k}{the number of individuals in a given frequency
  class}
}
\value{
  Alpha summed over all k
}
\description{
  calculate the alpha function using log factorials from
  Colwell 2012 see eqn 4
}

