\docType{data}
\name{costaRicaTrees}
\alias{costaRicaTrees}
\title{Costa Rican tree data}
\description{
  dataset of costa rican trees from table 4 of Colwell et
  al 2012.  Data is in three vectors, old_growth,
  older_growth, new_growth
}
\author{
  R Colwell \email{colwell at unconn.edu}
}
\references{
  J Plant Ecol (2012) 5 (1): 3-21.
  \url{http://dx.doi.org/10.1093/jpe/rtr044}
}
\keyword{datasets}

