\name{vec_rare}
\alias{vec_rare}
\title{Individual base rarefaction}
\usage{
  vec_rare(spp)
}
\arguments{
  \item{spp}{a vector of species abundances}
}
\value{
  interpolated values for all possible individual
  subsamples
}
\description{
  calculated individual based rarefaction for a single site
}

